Major improvements to hashing.
We weren't detecting hash collisions when building up the graph, only in the final (all_reachable=true) graphs. This was causing us in practice to not explore all graphs. For instance, we used to get 3 graphs on size = 3. Now we get 8.

Include the length of the node's list in the hash so that we don't hash two different graphs the same way.

Fix the order of operands when calling hash_combine.

Bump the hash size to 64-bit.

Manually peel the loop where we create the graphs. We can safely add seeds of 1-edge graphs n0->nx for all x > 1.

Allow out_degree_spec to exceed the size of the graph. For instance, a graph with size = 1 but has out-degree-spec of 2 represents a single binary operator. We want to allow that.